✨ feat(client): add constants module for Vue injection keys and classes
This commit introduces a new constants module within the client package, which defines a set of injection keys for Vue's dependency injection system, as well as a collection of CSS class constants. These constants will be used throughout the client application to standardize the way components interact with each other and to maintain a consistent naming convention for CSS classes. The addition of `TRUST_ORIGINS` array provides a centralized list of trusted domains, which can be used for security checks or cross-origin resource sharing policies.