✨ feat(iframe-left.vue): add new layout component for iframe with scaling feature
This commit introduces a new Vue component layout that includes an iframe on the left side of the screen, which can be scaled according to a provided scale property. The layout also supports a slot for additional content on the right side. This feature allows for more flexible UI design, particularly when embedding external content that needs to be resized.